requirements.txt
albumentations==2.0.8
matplotlib==3.10.8
opencv_python==4.12.0.88
Pillow==12.1.0
PyYAML==6.0.3
tqdm==4.67.1
ultralytics==8.3.233
torch
torchvision
numpy
supervision
tensorflow